Breakthrough, cards update again.

+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/tracing/card/TracingCardState.kt
@@ -17,7 +17,7 @@ data class TracingCardState(
     override val riskLevelLastSuccessfulCalculation: Int,
     override val matchedKeyCount: Int,
     override val daysSinceLastExposure: Int,
-    override val activeTracingDaysInRetentionPeriod: Int,
+    override val activeTracingDaysInRetentionPeriod: Long,
     override val lastTimeDiagnosisKeysFetched: Date?,
     override val isBackgroundJobEnabled: Boolean,
     override val isManualKeyRetrievalEnabled: Boolean,
