+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/tracing/details/TracingDetailsState.kt
@@ -14,7 +14,7 @@ data class TracingDetailsState(
     override val riskLevelLastSuccessfulCalculation: Int,
     override val matchedKeyCount: Int,
     override val daysSinceLastExposure: Int,
-    override val activeTracingDaysInRetentionPeriod: Int,
+    override val activeTracingDaysInRetentionPeriod: Long,
     override val lastTimeDiagnosisKeysFetched: Date?,
     override val isBackgroundJobEnabled: Boolean,
     override val isManualKeyRetrievalEnabled: Boolean,
